Reuse profiles/seccomp package
